Splitting hpp and cpp files properly
